+++ follow-order-followcom/src/main/java/net/maku/followcom/vo/FollowUploadTraderUserVO.java
@@ -34,13 +34,13 @@ public class FollowUploadTraderUserVO implements Serializable {
 	private Integer status;
 
 	@Schema(description = "上传数据数量")
-	private Long uploadTotal;
+	private Integer uploadTotal;
 
 	@Schema(description = "成功数量")
-	private Long successCount;
+	private Integer successCount;
 
 	@Schema(description = "失败数量")
-	private Long failureCount;
+	private Integer failureCount;
 
 	@Schema(description = "类型 0：新增账号 1：修改密码")
 	private Integer type;
